+++ src/main/app/case/models/task-list/index.ts
@@ -1,4 +1,4 @@
-export * from './enum/taskStatus';
+export * from './taskStatus';
 export * from './subTask';
 export * from './task';
 export * from './taskList';
