+++ src/main/app/case/models/task-list/taskStatus.ts
@@ -1,0 +1,4 @@
+export enum TaskStatus {
+  Complete = 'Complete',
+  NotStarted = 'NotStarted'
+}
