--- src/main/app/case/types/task-list/enum/taskStatus.ts
@@ -1,4 +1,0 @@
-export enum TaskStatus {
-  Complete = 'taskList:taskStatus.complete',
-  NotStarted = 'taskList:taskStatus.notStarted'
-}
